Refactor string concat fixes to be used in headless mode.
